LearnJava

+++ Main.java
@@ -2,7 +2,7 @@
 public class Main{
     public static void main(String[] args){
 
-        byte x = 130;
+        int x = 130;
 
         System.out.println("My number is: "+x);
     }
